+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/exception/UnsupportedRepositoryCategoryException.java
@@ -8,26 +8,11 @@ package com.dwarfeng.judge.stack.exception;
  */
 public class UnsupportedRepositoryCategoryException extends RepositoryException {
 
-    private static final long serialVersionUID = 7210479504024403804L;
+    private static final long serialVersionUID = 4035320694320450045L;
 
     private final String category;
 
     public UnsupportedRepositoryCategoryException(String category) {
-        this.category = category;
-    }
-
-    public UnsupportedRepositoryCategoryException(String message, Throwable cause, String category) {
-        super(message, cause);
-        this.category = category;
-    }
-
-    public UnsupportedRepositoryCategoryException(String message, String category) {
-        super(message);
-        this.category = category;
-    }
-
-    public UnsupportedRepositoryCategoryException(Throwable cause, String category) {
-        super(cause);
         this.category = category;
     }